修复类recordmanager的bug，现在不会出现奇怪错误

+++ ddf/RecordBinary.h
@@ -7,7 +7,7 @@
 struct RecordBinary
 {
     uchar* ptr;
-    size_t size;
+    ushort size;
 };
 
 #endif // _RECORD_BINARY_H
